edit profile, delete post, notifications

+++ lib/widgets/post_card.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:instagram_clone/providers/user_provider.dart';
 import 'package:instagram_clone/resources/firestore_methods.dart';
 import 'package:instagram_clone/screens/comment_screen.dart';
+import 'package:instagram_clone/screens/profile_screen.dart';
 import 'package:instagram_clone/utils/colors.dart';
 import 'package:instagram_clone/utils/utils.dart';
 import 'package:instagram_clone/widgets/like_animantion.dart';
@@ -50,7 +51,8 @@ class _PostCardState extends State<PostCard> {
 
   @override
   Widget build(BuildContext context) {
-    final User user = Provider.of<UserProvider>(context).getUser!;
+    final User user =
+        Provider.of<UserProvider>(context, listen: false).getUser!;
     final width = MediaQuery.of(context).size.width;
     return Container(
       decoration: BoxDecoration(
@@ -71,9 +73,19 @@ class _PostCardState extends State<PostCard> {
             ).copyWith(right: 0),
             child: Row(
               children: [
-                CircleAvatar(
-                  radius: 16,
-                  backgroundImage: NetworkImage(widget.snap['profImg']),
+                InkWell(
+                  onTap: () {
+                    Navigator.of(context).push(
+                      MaterialPageRoute(
+                        builder: (context) =>
+                            ProfileScreen(uid: widget.snap['uid']),
+                      ),
+                    );
+                  },
+                  child: CircleAvatar(
+                    radius: 16,
+                    backgroundImage: NetworkImage(widget.snap['profImg']),
+                  ),
                 ),
                 Expanded(
                   child: Padding(
@@ -82,10 +94,20 @@ class _PostCardState extends State<PostCard> {
                       mainAxisSize: MainAxisSize.min,
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        Text(
-                          widget.snap['username'],
-                          style: const TextStyle(
-                            fontWeight: FontWeight.bold,
+                        InkWell(
+                          onTap: () {
+                            Navigator.of(context).push(
+                              MaterialPageRoute(
+                                builder: (context) =>
+                                    ProfileScreen(uid: widget.snap['uid']),
+                              ),
+                            );
+                          },
+                          child: Text(
+                            widget.snap['username'],
+                            style: const TextStyle(
+                              fontWeight: FontWeight.bold,
+                            ),
                           ),
                         ),
                       ],
@@ -103,16 +125,24 @@ class _PostCardState extends State<PostCard> {
                                 ),
                                 shrinkWrap: true,
                                 children: [
-                                  'Delete',
+                                  widget.snap['uid'] == user.uid
+                                      ? 'Delete'
+                                      : 'Report',
                                 ]
                                     .map(
                                       (e) => InkWell(
                                         onTap: () async {
-                                          String res = await FirestoreMethods()
-                                              .deletePost(
-                                                  widget.snap['postId']);
-                                          Navigator.of(context).pop();
-                                          showSnackBar(res, context);
+                                          if (e == 'Delete') {
+                                            String res =
+                                                await FirestoreMethods()
+                                                    .deletePost(
+                                                        widget.snap['postId']);
+                                            Navigator.of(context).pop();
+                                            showSnackBar(res, context);
+                                          } else if (e == 'Report') {
+                                            Navigator.of(context).pop();
+                                            showSnackBar('Reported', context);
+                                          }
                                         },
                                         child: Container(
                                           padding: const EdgeInsets.symmetric(
